@@ -1,5 +1,4 @@
 {
-  "version": "2.1.12",
   "name": "menatwork/syncctopro",
   "description": "Synchronize multiple contao installations with each other",
   "keywords": [
@@ -34,8 +33,9 @@
     "source": "https://github.com/menatwork/syncCto"
   },
   "require": {
-    "php": ">=7.0",
-    "menatwork/synccto": "^4.0.0"
+    "php": ">=8.0",
+    "ext-zlib": "*",
+    "menatwork/synccto": "^5.0.0"
   },
   "autoload": {
     "psr-4": {
